+++ docs/JavaDocs/duke/Parser.html
@@ -123,7 +123,7 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 <li>java.lang.Object</li>
 <li>
 <ul class="inheritance">
-<li>duke.Parser</li>
+<li>duke.parser.Parser</li>
 </ul>
 </li>
 </ul>
@@ -176,7 +176,7 @@ extends java.lang.Object</pre>
 </tr>
 <tr id="i0" class="altColor">
 <td class="colFirst"><code>static void</code></td>
-<th class="colSecond" scope="row"><code><span class="memberNameLink"><a href="#parseCommands(java.lang.String,duke.TaskList,duke.Storage)">parseCommands</a></span>&#8203;(java.lang.String&nbsp;nextLine,
+<th class="colSecond" scope="row"><code><span class="memberNameLink"><a href="#parseCommands(java.lang.String,duke.tasks.TaskList,duke.dataFile.Storage)">parseCommands</a></span>&#8203;(java.lang.String&nbsp;nextLine,
              <a href="TaskList.html" title="class in duke">TaskList</a>&nbsp;tasks,
              <a href="Storage.html" title="class in duke">Storage</a>&nbsp;storage)</code></th>
 <td class="colLast">
@@ -226,7 +226,7 @@ extends java.lang.Object</pre>
 <!--   -->
 </a>
 <h3>Method Detail</h3>
-<a id="parseCommands(java.lang.String,duke.TaskList,duke.Storage)">
+<a id="parseCommands(java.lang.String,duke.tasks.TaskList,duke.dataFile.Storage)">
 <!--   -->
 </a>
 <ul class="blockListLast">
