+++ docs/JavaDocs/duke/Storage.html
@@ -123,7 +123,7 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 <li>java.lang.Object</li>
 <li>
 <ul class="inheritance">
-<li>duke.Storage</li>
+<li>duke.dataFile.Storage</li>
 </ul>
 </li>
 </ul>
@@ -183,7 +183,7 @@ extends java.lang.Object</pre>
 </tr>
 <tr id="i1" class="rowColor">
 <td class="colFirst"><code>void</code></td>
-<th class="colSecond" scope="row"><code><span class="memberNameLink"><a href="#write(duke.TaskList)">write</a></span>&#8203;(<a href="TaskList.html" title="class in duke">TaskList</a>&nbsp;tasks)</code></th>
+<th class="colSecond" scope="row"><code><span class="memberNameLink"><a href="#write(duke.tasks.TaskList)">write</a></span>&#8203;(<a href="TaskList.html" title="class in duke">TaskList</a>&nbsp;tasks)</code></th>
 <td class="colLast">
 <div class="block">Writes the current list of task into the storage file.</div>
 </td>
@@ -238,30 +238,30 @@ extends java.lang.Object</pre>
 <li class="blockList">
 <h4>load</h4>
 <pre class="methodSignature">public&nbsp;java.util.ArrayList&lt;<a href="Task.html" title="class in duke">Task</a>&gt;&nbsp;load()
-                               throws duke.DukeException</pre>
+                               throws duke.exceptions.DukeException</pre>
 <div class="block">Loads the taskList stored in the file.</div>
 <dl>
 <dt><span class="returnLabel">Returns:</span></dt>
 <dd>ArrayList representing the taskList.</dd>
 <dt><span class="throwsLabel">Throws:</span></dt>
-<dd><code>duke.DukeException</code> - Duke exception to let the user know what went wrong.</dd>
+<dd><code>duke.exceptions.DukeException</code> - Duke exception to let the user know what went wrong.</dd>
 </dl>
 </li>
 </ul>
-<a id="write(duke.TaskList)">
+<a id="write(duke.tasks.TaskList)">
 <!--   -->
 </a>
 <ul class="blockListLast">
 <li class="blockList">
 <h4>write</h4>
 <pre class="methodSignature">public&nbsp;void&nbsp;write&#8203;(<a href="TaskList.html" title="class in duke">TaskList</a>&nbsp;tasks)
-           throws duke.DukeException</pre>
+           throws duke.exceptions.DukeException</pre>
 <div class="block">Writes the current list of task into the storage file.</div>
 <dl>
 <dt><span class="paramLabel">Parameters:</span></dt>
 <dd><code>tasks</code> - list of task to be written.</dd>
 <dt><span class="throwsLabel">Throws:</span></dt>
-<dd><code>duke.DukeException</code> - Lets the user know what went wrong.</dd>
+<dd><code>duke.exceptions.DukeException</code> - Lets the user know what went wrong.</dd>
 </dl>
 </li>
 </ul>
