+++ docs/JavaDocs/duke/Task.html
@@ -123,7 +123,7 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 <li>java.lang.Object</li>
 <li>
 <ul class="inheritance">
-<li>duke.Task</li>
+<li>duke.tasks.Task</li>
 </ul>
 </li>
 </ul>
@@ -140,7 +140,7 @@ extends java.lang.Object
 implements java.io.Serializable</pre>
 <dl>
 <dt><span class="seeLabel">See Also:</span></dt>
-<dd><a href="../serialized-form.html#duke.Task">Serialized Form</a></dd>
+<dd><a href="../serialized-form.html#duke.tasks.Task">Serialized Form</a></dd>
 </dl>
 </li>
 </ul>
