+++ docs/JavaDocs/duke/Ui.html
@@ -123,7 +123,7 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 <li>java.lang.Object</li>
 <li>
 <ul class="inheritance">
-<li>duke.Ui</li>
+<li>duke.ui.Ui</li>
 </ul>
 </li>
 </ul>
@@ -183,7 +183,7 @@ extends java.lang.Object</pre>
 </tr>
 <tr id="i1" class="rowColor">
 <td class="colFirst"><code>static void</code></td>
-<th class="colSecond" scope="row"><code><span class="memberNameLink"><a href="#deleteItem(int,duke.Task)">deleteItem</a></span>&#8203;(int&nbsp;taskSize,
+<th class="colSecond" scope="row"><code><span class="memberNameLink"><a href="#deleteItem(int,duke.tasks.Task)">deleteItem</a></span>&#8203;(int&nbsp;taskSize,
           <a href="Task.html" title="class in duke">Task</a>&nbsp;deletedTask)</code></th>
 <td class="colLast">
 <div class="block">Prints after item has been deleted.</div>
@@ -198,7 +198,7 @@ extends java.lang.Object</pre>
 </tr>
 <tr id="i3" class="rowColor">
 <td class="colFirst"><code>static void</code></td>
-<th class="colSecond" scope="row"><code><span class="memberNameLink"><a href="#printItems(duke.TaskList)">printItems</a></span>&#8203;(<a href="TaskList.html" title="class in duke">TaskList</a>&nbsp;tasks)</code></th>
+<th class="colSecond" scope="row"><code><span class="memberNameLink"><a href="#printItems(duke.tasks.TaskList)">printItems</a></span>&#8203;(<a href="TaskList.html" title="class in duke">TaskList</a>&nbsp;tasks)</code></th>
 <td class="colLast">
 <div class="block">Print the items from the list of task.</div>
 </td>
@@ -287,7 +287,7 @@ extends java.lang.Object</pre>
 <div class="block">Says goodbye.</div>
 </li>
 </ul>
-<a id="printItems(duke.TaskList)">
+<a id="printItems(duke.tasks.TaskList)">
 <!--   -->
 </a>
 <ul class="blockList">
@@ -301,7 +301,7 @@ extends java.lang.Object</pre>
 </dl>
 </li>
 </ul>
-<a id="deleteItem(int,duke.Task)">
+<a id="deleteItem(int,duke.tasks.Task)">
 <!--   -->
 </a>
 <ul class="blockList">
