+++ docs/JavaDocs/index-files/index-1.html
@@ -97,7 +97,7 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 </a>
 <h2 class="title">A</h2>
 <dl>
-<dt><span class="memberNameLink"><a href="../duke/TaskList.html#addItem(duke.Parser.taskType,java.lang.String)">addItem(Parser.taskType, String)</a></span> - Method in class duke.<a href="../duke/TaskList.html" title="class in duke">TaskList</a></dt>
+<dt><span class="memberNameLink"><a href="../duke/TaskList.html#addItem(duke.parser.Parser.taskType,java.lang.String)">addItem(Parser.taskType, String)</a></span> - Method in class duke.<a href="../duke/TaskList.html" title="class in duke">TaskList</a></dt>
 <dd>
 <div class="block">Adds item to the list of tasks.</div>
 </dd>
