+++ docs/JavaDocs/index-files/index-10.html
@@ -97,7 +97,7 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 </a>
 <h2 class="title">P</h2>
 <dl>
-<dt><span class="memberNameLink"><a href="../duke/Parser.html#parseCommands(java.lang.String,duke.TaskList,duke.Storage)">parseCommands(String, TaskList, Storage)</a></span> - Static method in class duke.<a href="../duke/Parser.html" title="class in duke">Parser</a></dt>
+<dt><span class="memberNameLink"><a href="../duke/Parser.html#parseCommands(java.lang.String,duke.tasks.TaskList,duke.dataFile.Storage)">parseCommands(String, TaskList, Storage)</a></span> - Static method in class duke.<a href="../duke/Parser.html" title="class in duke">Parser</a></dt>
 <dd>
 <div class="block">Takes in commands from the user and does an action based on the command.</div>
 </dd>
@@ -105,7 +105,7 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 <dd>&nbsp;</dd>
 <dt><span class="memberNameLink"><a href="../duke/Parser.html#%3Cinit%3E()">Parser()</a></span> - Constructor for class duke.<a href="../duke/Parser.html" title="class in duke">Parser</a></dt>
 <dd>&nbsp;</dd>
-<dt><span class="memberNameLink"><a href="../duke/Ui.html#printItems(duke.TaskList)">printItems(TaskList)</a></span> - Static method in class duke.<a href="../duke/Ui.html" title="class in duke">Ui</a></dt>
+<dt><span class="memberNameLink"><a href="../duke/Ui.html#printItems(duke.tasks.TaskList)">printItems(TaskList)</a></span> - Static method in class duke.<a href="../duke/Ui.html" title="class in duke">Ui</a></dt>
 <dd>
 <div class="block">Print the items from the list of task.</div>
 </dd>
