+++ docs/JavaDocs/index-files/index-15.html
@@ -97,7 +97,7 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 </a>
 <h2 class="title">W</h2>
 <dl>
-<dt><span class="memberNameLink"><a href="../duke/Storage.html#write(duke.TaskList)">write(TaskList)</a></span> - Method in class duke.<a href="../duke/Storage.html" title="class in duke">Storage</a></dt>
+<dt><span class="memberNameLink"><a href="../duke/Storage.html#write(duke.tasks.TaskList)">write(TaskList)</a></span> - Method in class duke.<a href="../duke/Storage.html" title="class in duke">Storage</a></dt>
 <dd>
 <div class="block">Writes the current list of task into the storage file.</div>
 </dd>
