+++ docs/JavaDocs/index-files/index-4.html
@@ -97,7 +97,7 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 </a>
 <h2 class="title">D</h2>
 <dl>
-<dt><span class="memberNameLink"><a href="../duke/Ui.html#deleteItem(int,duke.Task)">deleteItem(int, Task)</a></span> - Static method in class duke.<a href="../duke/Ui.html" title="class in duke">Ui</a></dt>
+<dt><span class="memberNameLink"><a href="../duke/Ui.html#deleteItem(int,duke.tasks.Task)">deleteItem(int, Task)</a></span> - Static method in class duke.<a href="../duke/Ui.html" title="class in duke">Ui</a></dt>
 <dd>
 <div class="block">Prints after item has been deleted.</div>
 </dd>
