+++ docs/JavaDocs/index-files/index-9.html
@@ -105,7 +105,7 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 <dd>
 <div class="block">Indicates the task as marked.</div>
 </dd>
-<dt><span class="memberNameLink"><a href="../duke/TaskList.html#modifyStatus(duke.Parser.modifyStatus,java.lang.String)">modifyStatus(Parser.modifyStatus, String)</a></span> - Method in class duke.<a href="../duke/TaskList.html" title="class in duke">TaskList</a></dt>
+<dt><span class="memberNameLink"><a href="../duke/TaskList.html#modifyStatus(duke.parser.Parser.modifyStatus,java.lang.String)">modifyStatus(Parser.modifyStatus, String)</a></span> - Method in class duke.<a href="../duke/TaskList.html" title="class in duke">TaskList</a></dt>
 <dd>
 <div class="block">Marks/Un marks the task.</div>
 </dd>
