+++ docs/JavaDocs/serialized-form.html
@@ -101,10 +101,10 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 <section>
 <h2 title="Package">Package&nbsp;duke</h2>
 <ul class="blockList">
-<li class="blockList"><a id="duke.Task">
+<li class="blockList"><a id="duke.tasks.Task">
 <!--   -->
 </a>
-<h3>Class <a href="duke/Task.html" title="class in duke">duke.Task</a> extends java.lang.Object implements Serializable</h3>
+<h3>Class <a href="duke/Task.html" title="class in duke">duke.tasks.Task</a> extends java.lang.Object implements Serializable</h3>
 <ul class="blockList">
 <li class="blockList">
 <h3>Serialized Fields</h3>
@@ -121,10 +121,10 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 </li>
 </ul>
 </li>
-<li class="blockList"><a id="duke.TaskList">
+<li class="blockList"><a id="duke.tasks.TaskList">
 <!--   -->
 </a>
-<h3>Class <a href="duke/TaskList.html" title="class in duke">duke.TaskList</a> extends java.lang.Object implements Serializable</h3>
+<h3>Class <a href="duke/TaskList.html" title="class in duke">duke.tasks.TaskList</a> extends java.lang.Object implements Serializable</h3>
 <ul class="blockList">
 <li class="blockList">
 <h3>Serialized Fields</h3>
