+++ src/main/java/duke/dataFile/Storage.java
@@ -1,4 +1,9 @@
-package duke;
+package duke.dataFile;
+
+import duke.exceptions.DukeException;
+import duke.tasks.Task;
+import duke.tasks.TaskList;
+import duke.ui.Ui;
 
 import java.io.File;
 import java.io.FileInputStream;
